Corrections/TP2/requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.9.3
networkx==3.3
numpy==2.2.0
pandas==2.2.3
Requests==2.32.3
scikit_learn==1.5.2
seaborn==0.13.2
st_annotated_text==4.0.1
streamlit==1.40.1
tqdm==4.66.5
transformers==4.44.2
Unidecode==1.3.8
